+++ include/teqi/canvas/utility.h
@@ -11,24 +11,39 @@ namespace tq {
 ////////////////////////////////////////////////////////////////////////////////
 inline GLuint createGeometryBuffer(
     const std::vector<glm::vec3>& vertices,
+    const std::vector<glm::vec2>& coords,
     const std::vector<uint32_t>& indices,
     GLuint& vao,
     GLuint& vbo,
     GLuint& ibo,
     const GLuint usageHint = GL_STATIC_DRAW )
 {
-    const size_t vboSize = vertices.size() * sizeof( glm::vec3 );
+    const size_t vertSize = vertices.size() * sizeof( glm::vec3 );
+    const size_t coordSize = coords.size() * sizeof( glm::vec2 );
+    const size_t vboSize = vertSize + coordSize;
     const size_t iboSize = indices.size() * sizeof( uint32_t );
+
+    //  vertex array
 
     if ( ! vao ) glGenVertexArrays( 1, &vao );
     glBindVertexArray( vao );
 
+    //  vertex buffer
+
     if ( ! vbo ) glGenBuffers( 1, &vbo );
     glBindBuffer( GL_ARRAY_BUFFER, vbo );
-    glBufferData( GL_ARRAY_BUFFER, vboSize, vertices.data(), usageHint );
+    glBufferData( GL_ARRAY_BUFFER, vboSize, NULL, usageHint );
+
+    glBufferSubData( GL_ARRAY_BUFFER, 0, vertSize, vertices.data() );
+    glBufferSubData( GL_ARRAY_BUFFER, vertSize, coordSize, coords.data() );
 
     glVertexAttribPointer( 0, 3, GL_FLOAT, GL_FALSE, sizeof( glm::vec3 ), ( void* )( 0 ) );
+    glVertexAttribPointer( 1, 2, GL_FLOAT, GL_FALSE, sizeof( glm::vec2 ), ( void* )( vertSize ) );
+
     glEnableVertexAttribArray( 0 );
+    glEnableVertexAttribArray( 1 );
+
+    //  index buffer
 
     if ( ! ibo ) glGenBuffers( 1, &ibo );
     glBindBuffer( GL_ELEMENT_ARRAY_BUFFER, ibo );
@@ -43,12 +58,23 @@ inline GLuint createGeometryBuffer(
 
 
 ////////////////////////////////////////////////////////////////////////////////
-inline void drawQuad( const glm::vec2& pos, const glm::vec2& size, const float depth = 0.f )
+inline void drawQuad(
+    const glm::vec2& pos,
+    const glm::vec2& size,
+    const float depth = 0.f )
 {
     static GLuint vao = 0, vbo = 0, ibo = 0;
+
     static const std::vector<uint32_t> indices = {
         0, 1, 3,
         1, 2, 3
+    };
+
+    static const std::vector<glm::vec2> coords = {
+        { 0, 0 },
+        { 1, 0 },
+        { 1, 1 },
+        { 0, 1 }
     };
 
     const std::vector<glm::vec3> vertices = {
@@ -58,7 +84,41 @@ inline void drawQuad( const glm::vec2& pos, const glm::vec2& size, const float d
         { pos.x, pos.y + size.y, depth }
     };
 
-    createGeometryBuffer( vertices, indices, vao, vbo, ibo );
+    createGeometryBuffer( vertices, coords, indices, vao, vbo, ibo );
+
+    glBindVertexArray( vao );
+    glDrawElements( GL_TRIANGLES, 6, GL_UNSIGNED_INT, 0 );
+    glBindVertexArray( 0 );
+}
+
+
+////////////////////////////////////////////////////////////////////////////////
+inline void drawUnitQuad()
+{
+    static GLuint vao = 0, vbo = 0, ibo = 0;
+
+    static const std::vector<uint32_t> indices = {
+        0, 1, 3,
+        1, 2, 3
+    };
+
+    static const std::vector<glm::vec2> coords = {
+        { 0, 0 },
+        { 1, 0 },
+        { 1, 1 },
+        { 0, 1 }
+    };
+
+    static const std::vector<glm::vec3> vertices = {
+        {-1,-1, 0 },
+        { 1,-1, 0 },
+        { 1, 1, 0 },
+        {-1, 1, 0 }
+    };
+
+    if ( ! vao ) {
+        createGeometryBuffer( vertices, coords, indices, vao, vbo, ibo );
+    }
 
     glBindVertexArray( vao );
     glDrawElements( GL_TRIANGLES, 6, GL_UNSIGNED_INT, 0 );